Moved setup section down
Brought the setup section down from being at the top of the README. It was confusing being the first thing users read. It may be a good team and self-reminder, but there is little context to new users, nor an explanation of what the section is meant to help with. There still needs to be work done on this section, as it is not completely clear to me what it does and why Ropsten functions were mentioned.